% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/event.registry.R
\name{add.event.handler}
\alias{add.event.handler}
\title{add.event.handler}
\usage{
add.event.handler(registry, event, handler)
}
\arguments{
\item{registry}{EventRegistry}

\item{event}{String. Name of the Event}

\item{handler}{Function. The new Handler to add to the Event.}
}
\value{
Nothing good.
}
\description{
Add a Handler to an Event
}
\details{
A Handler is any function to be called when the event is triggered.
If the return value of the Handler has a "CatchEvent" attribute which is TRUE then the
event will be caught and not bubble to the next handler, and the "CatchEvent" attribute
will be stripped before returning the value to the triggering context.

If the Event does not yet exist an error is thrown.
}
\examples{
r <- new.event.registry()
add.event(r, "mouseclick")
add.event.handler(r, "mouseclick", function(x, y)  message("Mouse clicked at coordinates (", x, ", ", y, ")"))
trigger.event(r, "mouseclick", x = 30, y = 50)
}
\author{
Brad Friedman
}

