% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/event.registry.R
\name{add.event}
\alias{add.event}
\title{add.event}
\usage{
add.event(registry, event, overwrite = FALSE)
}
\arguments{
\item{registry}{EventRegistry}

\item{event}{String. Name for the Event.}

\item{overwrite}{Logical. If the Event already exists, should I overwrite it? If TRUE
then yes, without warning. If FALSE (default) then no, throw an error.}
}
\value{
Nothing good.
}
\description{
Add a new event to a registry
}
\details{
Add a new event to a registry. If an Event of that name already
exists then if \code{overwrite} is not set then an error is thrown, and
if \code{overwrite} is set then the contents of the old Event are simply replaced.
Use \code{add.handler} to add a handler to an existing Event.
}
\examples{
r <- new.event.registry()
has.event(r, "mouseclick")
add.event(r, "mouseclick")
has.event(r, "mouseclick")
}
\author{
Brad Friedman
}

