% Generated by roxygen2 (4.0.2): do not edit by hand
\name{analysis.link}
\alias{analysis.link}
\title{analysis.link}
\usage{
analysis.link(page, params = list(), app.base.url, width = 9, height = 7,
  device = "svg", include.plot.params = TRUE)
}
\arguments{
\item{page}{Name of page}

\item{params}{List of parameter values (as R objects---this function will encode them). Default: \code{list()} (no parameters).}

\item{app.base.url}{Base URL for application. This is usually the prefix in which the app landing HTML page is found.}

\item{width}{Width parameter for graphics devices. Normally this is in inches, although it depends on exactly
how you set up your application. Default: 9. Ignored if \code{include.plot.params = FALSE}}

\item{height}{Height parameter for graphics devices. Normally this is in inches, although it depends on exactly
how you set up your application. Default: 7. Ignored if \code{include.plot.params = FALSE}}

\item{device}{Device to use for plotting. Default: "svg". (This is going to be sent to the server, not run
in the same process as this function. It would be ignored if the page was built with \code{no.plot = TRUE}.)
Ignored if \code{include.plot.params = FALSE}}

\item{include.plot.params}{Boolean, default TRUE. If TRUE then include the width, height, device and
textarea_wrap parameters in the URL.
in the URL. Otherwise omit them.}
}
\value{
URL
}
\description{
Build a URL to run an analysis on the server
}
\details{
Unlike \code{\link{analysis.page.link}} this is not a URL to open in the web browser, but rather the kind of URL used internally by the
front end for its AJAX request to the server to perform an analysis, or retrieve the analyses as web services.
}
\author{
Brad Friedman
}

