% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/current.R
\name{analysis.page.of.current.app}
\alias{analysis.page.of.current.app}
\title{Retrieve an Analysis page from the current app}
\usage{
analysis.page.of.current.app(page)
}
\arguments{
\item{page}{String, name of desired page.}
}
\value{
AnalysisPage, or NULl
}
\description{
If the current app (as returned by (\code{\link{current.app}})
has a page of the given name then it is returned. If
the current app can't be found, or if it does not have such a
page, then NULL is returned.
}
\author{
Brad Friedman
}

