% Generated by roxygen2 (4.0.2): do not edit by hand
\name{annotate.analysis.page.svg}
\alias{annotate.analysis.page.svg}
\title{annotate.analysis.page.svg}
\usage{
annotate.analysis.page.svg(svg.filename, x, y, ids, group.lengths = length(x),
  class.name = "plot-point", start = 0, uniquify.ids.suffix = NULL,
  verbose = FALSE)
}
\arguments{
\item{svg.filename}{Path to SVG file}

\item{x}{Vector of x values}

\item{y}{Vector of y values (same length as x)}

\item{ids}{Vector of ids (same length as x)}

\item{group.lengths}{Positive integer vector summing to \code{x} (an ordered partition of \code{length(x)})
giving length of contiguous groups of elements. Each integer must be at least 3. The search will be for
contiguous elements, then continue where the last one left off, possibly after a gap. Default: \code{length(x)},
look for a single contiguous block.}

\item{class.name}{Class name to apply to points (default: "plot-point")}

\item{start}{This is a 0-based integer. It is an offset of where to start
looking for the plot elements. Default is 0, to start at the beginning. The meaning
of the index after that is not exposed to the interface, however this function returns numbers that
you can then use to "continue looking where you left off".}

\item{uniquify.ids.suffix}{NULL or a string. If NULL then do not modify the
identifiers in the SVG file. If a string, then call \code{\link{uniquify.ids.in.svg.files}}
to modify the "glyph" and "clip" identifiers, using this word as the \code{suffixes} parameter).}

\item{verbose}{Logical. Default, FALSE. Passed through (as integer) to C++}
}
\value{
Integer, the "next start" position, or where to start looking to continue after this
stretch. (Invisibly). If no match was found then returns NULL.
}
\description{
Annotate an AnalysisPage SVG plot
}
\details{
The plot points are found by looking for a sequence of points whose x and y
coordinates correlation >= 0.999 with the query x and y vectors.

Then each is tagged with class.name, and with the corresponding ID from the vector.

The file is overwritten.
}
\author{
Brad Friedman
}

