% Generated by roxygen2 (4.0.2): do not edit by hand
\name{apache.httpd.conf}
\alias{apache.httpd.conf}
\title{apache.httpd.conf}
\usage{
apache.httpd.conf(driver.path, app.location, config.js.path,
  front.end.dir = system.file("htdocs/client/dist-aps", package =
  "AnalysisPageServer"), mod.R.path, skip.checks = FALSE)
}
\arguments{
\item{driver.path}{Path to driver. Must contain call to add.handlers.to.global().}

\item{app.location}{Location from which app will be deployed (e.g. "/myapp" to make the URL
start "http://myserver.com/myapp").}

\item{config.js.path}{Path to modified \code{config.js} file. This would have been generated
with a call to \code{\link{config.js}}.}

\item{front.end.dir}{Path to front end directory content. An alias will be set up to
serve it directly from this location. Default is from the installed R package (found via
\code{system.file}.}

\item{mod.R.path}{Path to mod_R.so}

\item{skip.checks}{Boolean, default FALSE. If TRUE then don't check for file existence or for
the presence of "add.handlers.to.global" in the startup script.}
}
\value{
Charvec
}
\description{
Generate httpd.conf file for RApache deployment
}
\details{
Generate httpd.conf file for RApache deployment. This returns a charvec of lines of the files.
You still have to call \code{writeLines}.
See the \code{ApacheDeployment} vignette for more information.
}
\author{
Brad Friedman
}

