% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom.html.R
\name{aps.dataset.divs}
\alias{aps.dataset.divs}
\title{aps.dataset.divs}
\usage{
aps.dataset.divs(paths.list, show.sidebar = rep(TRUE, length(paths.list)),
  show.table = rep(TRUE, length(paths.list)), num.table.rows = 10,
  extra.html.class = rep(list(character()), length(paths.list)),
  extra.div.attr = rep(list(NULL), length(paths.list)))
}
\arguments{
\item{paths.list}{List of lists. The outer list corresponds to data sets
and the inner lists have names in \code{$plot} and \code{$data}, giving
relative paths to the SVG and JSON files (OK to omit one). Or a list
with a \code{$paths.list} element, which would be used (this lets you
pass the return value of \code{static.analysis.page} directly to this function).}

\item{show.sidebar}{Logical vector of same length as \code{paths.list} to
pass through corresponding elements to \code{\link{aps.one.dataset.div}}.
Default: all TRUE.}

\item{show.table}{Logical vector of same length as \code{paths.list} to
pass through corresponding elements to \code{\link{aps.one.dataset.div}}.
Default: all TRUE.}

\item{num.table.rows}{Number of table rows to show. Default: 10.
Recycled to \code{length(paths.list)}.}

\item{extra.html.class}{List (of charvecs) of same length as \code{paths.list} to
pass through corresponding elements to \code{\link{aps.one.dataset.div}}.
Default: All empty charvec.}

\item{extra.div.attr}{List (of named charvecs or NULLs) of same length as \code{paths.list} to
pass through corresponding elements to \code{\link{aps.one.dataset.div}}.
Default: all NULL.}
}
\value{
Charvec of HTML divs corresponding to datasets in \code{paths.list}.
}
\description{
Generate HTML for multiple DIV elements corresponding to a paths list
}
\details{
This function is meant to work with the return value of \code{\link{static.analysis.page}}
That function returns an object with a \code{$paths.list} element which
contains the relative paths to each of the plots and datasets. You pass that
through as the first argument to this function and it will make divs corresponding
to those plots. The other arguments are either vectors or lists of corresponding
lengths to pass through to \code{\link{aps.one.dataset.div}}.
}
\author{
Brad Friedman
}

