% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom.html.R
\name{aps.one.dataset.div}
\alias{aps.one.dataset.div}
\title{aps.one.dataset.div}
\usage{
aps.one.dataset.div(svg.path = NULL, data.path = NULL,
  show.sidebar = TRUE, show.table = TRUE, allow.zoom = TRUE,
  plot.height = NULL, div.width = NULL, style = NULL,
  num.table.rows = 10, extra.html.class = character(),
  extra.div.attr = NULL)
}
\arguments{
\item{svg.path}{Path (could be relative to index.html) to (annotated) SVG file. NULL
to only have data table and no picture.}

\item{data.path}{Path (could be relative to index.html) to JSON file containing data set.
NULL to only have SVG and no table}

\item{show.sidebar}{Boolean. If TRUE (default) then show sidebar. If FALSE then omit it.}

\item{show.table}{Boolean. If TRUE (default) then show sidebar. If FALSE then omit it.}

\item{allow.zoom}{If TRUE (default) then allow zooming and panning. IF FALSE then
do not allow it.}

\item{plot.height}{If NULL (default) then do not specify 'data-plot-height' attribute.
Otherwise, use this number as 'data-plot-height' attribute, which will specify
the plot height (in pixels)}

\item{div.width}{If NULL (default) then do not specify div width in style.
Otherwise, supply a valid CSS width (e.g. "200px" or "60%")
and this will be rolled into the inline-style}

\item{style}{String specifying inline style of this div or NULL (default).
If NULL then and \code{div.width}
is also NULL then do not specfiy any inline style. If NULL and \code{div.width} is
non-NULL then create a centered div of \code{div.width} pixels wide with
\code{style="width:100px; margin:0 auto"} (or whatever div.width is, instead of "100px").
If non-NULL then use the string directly as the style attribute of the div.}

\item{num.table.rows}{Number of table rows to show. Default: 10}

\item{extra.html.class}{Thesee are extra classes to add to the div. This could be used for
whatever extended purpose you want, like extra styling or logic. Should be an unnamed charvec.
Default is \code{character()}, just use the basic required for APS.}

\item{extra.div.attr}{These are extra attributes to add to the div. For example you could
add an \code{id} attribute. It should be a named charvec, or NULL (default) to not anything
extra beyond that required for APS.}
}
\value{
HTML string
}
\description{
Create HTML for a div element to contain one AnalysisPageServer data set
}
\details{
Create HTML for a div element to contain one AnalysisPageServer data set.
This function does not created, modify, or even check for existance of
the SVG and JSON files. You provide paths and this function just includes
those paths, however awful, in the HTML returned.
}
\seealso{
\code{\link{aps.dataset.divs}}, a convenience wrapper for this
function to create multple divs at once.
}
\author{
Brad Friedman
}
