% Generated by roxygen2 (4.0.2): do not edit by hand
\name{aps.urlEncode}
\alias{aps.urlEncode}
\title{aps.urlEncode}
\usage{
aps.urlEncode(vec)
}
\arguments{
\item{vec}{Vector to encode. Will be coerced to character.}
}
\value{
charvec of same length, with encoded strings
}
\description{
urlEncode all the strings in a character vector
}
\details{
The urlEncode function supplied by RApache has a few behaviors I don't expect.
First, it requires only character arguments, so \code{urlEncode(1)} throws
an exception. Second, it dies on the empty string.

This function is a wrapper for RApache's \code{urlEncode} that handles all
these cases. It might later be replaced with another implementation, but
the interface will stay. (The "aps" stands for \code{AnalysisPageServer}.)
}
\examples{
aps.urlEncode(1)
  aps.urlEncode("")
  aps.urlEncode("foo/bar")
}
\author{
Brad Friedman - Regular
}

