% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory.R
\name{autosignal.on.bloated.memory}
\alias{autosignal.on.bloated.memory}
\title{autosignal.on.bloated.memory}
\usage{
autosignal.on.bloated.memory(events, signal = tools::SIGUSR1,
  pid = Sys.getpid(), logger = NULL, event.name = "BloatedMemory")
}
\arguments{
\item{events}{EventRegistry object}

\item{signal}{Signal to throw. Default is SIGUSR1 (defined in tools package). Normally this elicits similar behavior to SIGINT
however when running inside Apache it lets the worker process finish the current request
before killing itself.}

\item{pid}{Process ID to which the signal should be sent. Default: \code{Sys.getpid()}, the current process. (If you
provide something else it is no longer really an \code{autosignal}, just a \code{signal}).}

\item{logger}{log4r object, optional}

\item{event.name}{Name of event to listen for. Default: "BloatedMemory"}
}
\value{
Nothing
}
\description{
Send signal to self on BloatedMemory
}
\details{
Have the process send a signal to itself when BloatedMemory event is triggered.
}
\note{
This attaches a listener---it does not actaully do anything until the BloatedMemory event is triggered, if ever.
}
\author{
Brad Friedman
}
