% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory.R
\name{bind.memory.checker}
\alias{bind.memory.checker}
\title{bind.}
\usage{
bind.memory.checker(app, max.mb, app.event = "FinishAnalysis",
  memory.event = "BloatedMemory")
}
\arguments{
\item{app}{AnalysisPageRApacheApp}

\item{max.mb}{Memory threshold for triggering BloatedMemory, in Megabytes. Required.}

\item{app.event}{Name of existing event on which the new memory check listener should
be registered. Default: "FinishAnalysis".}

\item{memory.event}{Name of event to fire in case of memory usage above threshold.
Default: "BloatedMemory".}
}
\value{
Nothing of note.
}
\description{
Bind the memory checker to the FinishAnalysis event
}
\details{
You supply an AnalysisPageRApacheApp. It also has a FinishAnalysis
event. I add a new event (by default called BloatedMemory), and
also a listener for FinishAnalysis, which calls \code{\link{check.memory}}
each time \code{FinishAnalysis} is triggered.
(this will then trigger BloatedMemory if memory usage is above threshold).
Although the BloatedMemory event would now be triggered, unless a listener
is attached to *it* nothing special will happen.
(See \code{\link{autosignal.on.bloated.memory}} for this.)
}
\seealso{
\code{\link{check.memory}}, \code{\link{autosignal.on.bloated.memory}}
}
\author{
Brad Friedman
}
