% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/params.R
\name{bool.param}
\alias{bool.param}
\title{bool.param}
\usage{
bool.param(...)
}
\arguments{
\item{...}{Passed through to \code{\link{simple.param}}. This includes at least "name", and optionally
"label", "description" and "value" (which should be either \code{TRUE} or \code{FALSE}),
but not "type".}
}
\value{
An AnalysisPageParam
}
\description{
Build a boolean AnalysisPageParam
}
\details{
Build a boolean AnalysisPageParam. This is probably rendered as a checkbox. The value returned to
the server should be either JSON "true" (corresponding to checked) or JSON "false".

If you do not provide a value (or if you provide value="", which is what happens in the parent
constructor \code{simple.param} when you don't provide a value) then the default will be FALSE.
}
\examples{
show.ids <- bool.param("show_ids", label="Show IDs", description="Show sample IDs on the plot", value=TRUE)
}
\author{
Brad Friedman
}

