% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service.R
\name{build.service}
\alias{build.service}
\title{build.service}
\usage{
build.service(handler, param.set = default.service.paramset(handler),
  annotate.plot = FALSE, annotate.data.frame = FALSE, no.plot = TRUE,
  service = TRUE, skip.checks = TRUE, ...)
}
\arguments{
\item{handler}{The handler function to convert}

\item{param.set}{AnalysisPageParamSet for the handler. Default: \code{default.service.paramset(handler)}}

\item{annotate.plot}{Default: FALSE}

\item{annotate.data.frame}{Default: FALSE}

\item{no.plot}{Default: TRUE}

\item{service}{Default: TRUE}

\item{skip.checks}{Default: TRUE. This is passed through to \code{\link{new.analysis.page}} and should
normally not be modified. (we don't check services because they are not required to provide default
arguments.)}

\item{...}{Further arguments to pass to \code{\link{new.analysis.page}}}
}
\value{
\code{AnalysisPage}
}
\description{
Build an AnalysisPage service
}
\details{
Convert a functions into an
AnalysisPage service. An AnalysisPage service is an AnalysisPage with the service flag set.
That means that pages() will not return it, so it will not be directly available through the
front end. The reason to do this would normally be so that it can be used to populate
a combobox or other part of the website. In practice it also means that the return
value will probably not use the AnalysisPageDataNode system---it is free to return
some arbitrary JSON string or other text.

It is a wrapper for \code{\link[AnalysisPageServer]{new.analysis.page}}, and passes all of
its arguments through (but with different defaults now)
}
\examples{
poem.file <- system.file("examples/in-a-station-of-the-metro.html", package="AnalysisPageServer")
poem.html <- readLines(poem.file, warn = FALSE)
poem <- build.service(function()  {
  new.response(paste0(poem.html, "\\n"),
  content.type = "text/html")
}, name = "poem")
}
\author{
Brad Friedman
}
\seealso{
\code{\link[AnalysisPageServer]{new.analysis.page}}
}

