% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory.R
\name{check.memory}
\alias{check.memory}
\title{check.memory}
\usage{
check.memory(events, max.mb, logger = NULL, event.name = "BloatedMemory")
}
\arguments{
\item{events}{EventRegistry object}

\item{max.mb}{If the total "used" memory (Ncells + Vcells) is at least this much
(in Mb) then trigger the next handler.}

\item{logger}{log4r object. If non-NULL then memory usage is printed there with \code{info()}. Default: NULL, don't log.}

\item{event.name}{Name of the event to trigger. Default: "BloatedMemory".
It is passed \code{used.mb} and \code{max.mb} arguments}
}
\value{
Nothing
}
\description{
Check memory usage and trigger an event if it exceeds some threshold.
}
\details{
Call \code{gc()} to check memory,
possibly print the result to STDERR, then possibly
call the next handler you supply if too much memory is being
used.
}
\author{
Brad Friedman
}

