% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test.R
\name{check.signal}
\alias{check.signal}
\title{check.signal}
\usage{
check.signal(expr, signo, testname, no.signal = FALSE)
}
\arguments{
\item{expr}{The expression to evaluate}

\item{signo}{The signal number (consider using constants
like SIGUSR1 from the tools package).}

\item{testname}{Name for this test. Default is to build from
signo argument.}

\item{no.signal}{Logical, to invert the sense of the test.
Default, FALSE, means to test that the signal was delivered.
TRUE means to test that the signal was not delivered.}
}
\value{
The value of the evaluated expression, invisibly,
so you can do more testing if desired.
}
\description{
Check if an expression results in a signal being delivered
}
\details{
Check if an expression results in a signal being delivered.
The signal will be caught: you can safely deliver a signal
such as SIGUSR1 that would normally cause the process to die.
}
\author{
Brad Friedman - Regular
}

