% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rook.analysis.page.server.R
\name{checkPackageInstalled}
\alias{checkPackageInstalled}
\title{Checks if a given package is installed.}
\usage{
checkPackageInstalled(pkg, version = "0.0.0", required = FALSE)
}
\arguments{
\item{pkg}{A character string containing a package name.}

\item{version}{A minimum version number. Default: "0.0.0" (no version requirement)}

\item{required}{A boolean. The functions stops if set to \code{TRUE} and if the required package is not present. Default is \code{FALSE}.}
}
\value{
A boolean.
}
\description{
Checks if a given package is installed.
}
\examples{
checkPackageInstalled("AnalysisPageServer")
}
\author{
Cory Barr
}
