% Generated by roxygen2 (4.0.2): do not edit by hand
\name{checkRookForkForVignettes}
\alias{checkRookForkForVignettes}
\title{checkRookForkForVignettes}
\usage{
checkRookForkForVignettes(rookforkOK)
}
\arguments{
\item{rookforkOK}{Provide FALSE here if you want to simulate not
having valid installed copies of rook/fork, without actually having
to delete them. Normally you should not supply this argument}
}
\value{
TRUE if Rook (>= 1.1) and fork are available, otherwise FALSE.
}
\description{
Check availability of Rook and fork for vignettes
}
\details{
Check availability of Rook and fork for vignettes. This function is only
meant to be called at the top of the ExampleServers.Rmd and InteractiveApps.Rmd
vignettes. It checks that the user has Rook >= 1.1 and fork installed.
If not, then it emits a useful message about how to install it, and that
the vignette will not build with all features. It also writes functions
\code{kill.process}, \code{readLines} (yes! be careful!),
\code{startRookAnalysisPageServer} and \code{new.rook.analysis.page.app}
in the Global namespace
to avoid calling the real functuons and instead just emit a short
message that Rook/fork are not available. The message is also available
in the global variable \code{no.rook.fork.msg}.

Really, you shouldn't
use this function except if you are writing a new vignette.
}
\note{
Are you sure you really want to use this function? Probably not,
unless you are calling it from the top of a new vignette. Otherwise
your are really asking for trouble messing up your global namespace.
See Details.

Why are you still here? Didn't I tell you not to call this function?
}
\author{
Brad Friedman
}

