% Generated by roxygen2 (4.0.2): do not edit by hand
\name{clear.event.handlers}
\alias{clear.event.handlers}
\title{clear.event.handlers}
\usage{
clear.event.handlers(registry, event)
}
\arguments{
\item{registry}{EventRegistry}

\item{event}{String. Name of the Event.}
}
\value{
Nothing good.
}
\description{
Clear the Handlers list for one Event
}
\details{
Clear the Handlers list for one Event. Does not remove the Event from the EventRegistry.
}
\examples{
r <- new.event.registry()
add.event(r, "mouseclick")
add.event.handler(r, "mouseclick", function(x, y)  message("Mouse clicked at coordinates (", x, ", ", y, ")"))
trigger.event(r, "mouseclick", x = 30, y = 50)
clear.event.handlers(r, "mouseclick")
trigger.event(r, "mouseclick", x = 30, y = 50)
}
\author{
Brad Friedman
}

