% Generated by roxygen2 (4.0.2): do not edit by hand
\name{compound.param}
\alias{compound.param}
\title{compound.param}
\usage{
compound.param(..., children)
}
\arguments{
\item{...}{Passed through to \code{\link{simple.param}}. This includes at least "name", optionally "label" and "description".
but not "type" (which is set to "compound") or "value" (which is set to empty string, but ignored anyway, since each of the
contained parameters will have its own value).}

\item{children}{AnalysisPageParamSet An AnalysisPageParamSet representing all of the nested parameters}
}
\value{
AnalysisPageParam of type "compound"
}
\description{
Create a Compound AnalysisPageParam
}
\details{
A compound AnalysisPageParam is a single parameter that has multiple parts.
The parts are represented by an AnalysisPageParamSet, so could be arbitrarily
nested. The front end is responsible for wrapping up all of the values
in a JSON hash and passing in a single value.

This can be thought of as a way of building a hash out of other parameter types.
}
\examples{
plist <- list(simple.param(name="study"), simple.param(name="comp"), simple.param("feature.type", value="gene"))
  comp <- compound.param(name="comp", label="Comparison", children=param.set(plist))
}
\author{
Brad Friedman
}

