% Generated by roxygen2 (4.0.2): do not edit by hand
\name{config.js}
\alias{config.js}
\title{config.js}
\usage{
config.js(app.prefix = "/custom/RAPS", client.r.url = file.path(app.prefix,
  "R"), client.rest.url = "",
  template.file = system.file("config-template.js", package =
  "AnalysisPageServer"), static = FALSE, parameter.collection.url = if
  (static) "" else "params", page.collection.url = if (static) "" else
  "pages")
}
\arguments{
\item{app.prefix}{Prefix for the path to your application. This will be used as the value for the "history.root"
parameter in the Javascript file, and also to build the default \code{client.r.url}.}

\item{client.r.url}{Location of R resources. Default: \code{file.path(app.prefix, "R")}.}

\item{client.rest.url}{Location of Sloth REST resources. Default: \code{""}. This is a poorly documented feature
that most people should ignore.}

\item{template.file}{Path to template file for config.js. Default is taken from "inst" directory of AnalysisPageServer package.}

\item{static}{Boolean, default FALSE. Controls the default values for \code{parameter.collection.url}
and \code{page.collection.url}.}

\item{parameter.collection.url}{Default: If \code{static = TRUE} then "" else "params"}

\item{page.collection.url}{Default: If \code{static = TRUE} then "" else "pages"}
}
\value{
Charvec of Javascript
}
\description{
Build Javascript configuration
}
\details{
Build Javascript configuration. This function returns Javascript which can be used as the config.js file for
the front-end client.

The only reason to call this directly would be to set up specialized deployments.
}
\author{
Brad Friedman
}

