% Generated by roxygen2 (4.0.2): do not edit by hand
\name{copy.front.end}
\alias{copy.front.end}
\title{copy.front.end}
\usage{
copy.front.end(outdir, client.basedir = system.file("htdocs/client/dist-apss",
  package = "AnalysisPageServer"), ...)
}
\arguments{
\item{outdir}{Target directory. This directory will contain your index.html file.}

\item{client.basedir}{Path to client files. Default: \code{system.file("htdocs/client/dist-apss", package = "AnalysisPageServer")}.
Probably should not be modified except during development work on the client.}

\item{...}{Passed through to \code{file.copy}, such as \code{overwrite = TRUE}}
}
\value{
Whatever file.copy returns.
}
\description{
Copy the APS front end (HTML, CSS, JS, etc) to a web directory
}
\details{
This makes a copy of the complete APS \emph{static}
front end (HTML, CSS, JS, etc) to a web directory.
}
\examples{
message("See vignette embedding.html")
}
\author{
Brad Friedman
}

