% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/current.R
\name{current.app}
\alias{current.app}
\title{Return the currently running AnalysisPage app}
\usage{
current.app()
}
\value{
Current \code{AnalysisPageRApacheApp}, or NULL
if it can't be found.
}
\description{
Return the currently running AnalysisPage app. The way this is done
is to first try to chase up the call stack and find the first environment which is
an \code{AnalysisPageRApacheApp}, and return that. If that fails
then it looks for \code{app} in the GlobalEnv and returns that. IF that
also fails then it reutrns NULL.
}
\author{
Brad Friedman
}

