% Generated by roxygen2 (4.0.2): do not edit by hand
\name{custom.body.attr}
\alias{custom.body.attr}
\title{custom.body.attr}
\usage{
custom.body.attr()
}
\value{
Name charvec of attributes for body
}
\description{
Return custom attributes required for body element
}
\details{
This attribute must be included in the <body> element.
}
\author{
Brad Friedman
}

