% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom.html.R
\name{custom.html.headers}
\alias{custom.html.headers}
\title{custom.html.headers}
\usage{
custom.html.headers(libbase.prefix = "",
  viewport = "width=device-width, initial-scale=1.0",
  stylesheets = default.stylesheets(), ep.svg.styles = NULL)
}
\arguments{
\item{libbase.prefix}{Prefix where your shared CSS files will be located. Default "" will be
relative to the index.html file. Otherwise you'll need to end with a "/".}

\item{viewport}{Default: "width=device-width, initial-scale=1.0". This will be used in a \code{<meta name="viewport">} tag.}

\item{stylesheets}{Charvec of stylesheets to load. Default is \code{default.stylesheets()}.}

\item{ep.svg.styles}{ep-svg-styles stylesheet. Default: NULL.}
}
\value{
HTML string to be included in \code{<head>} section.
}
\description{
Generate HTML for custom headers to load AnalysisPageServer CSS and viewport
}
\details{
Generate HTML for custom headers to load AnalysisPageServer CSS and viewport.
To be honest I don't understand how all this works. The main thing is that it you put this stuff
up top, in the header section. The only argument you should consider touching is \code{libbase.prefix},
if you are going to put your common libraries in a shared area instead making a copy next to each
dataset.
}
\author{
Brad Friedman
}
