% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapache.R
\name{data.frame.to.json}
\alias{data.frame.to.json}
\title{data.frame.to.json}
\usage{
data.frame.to.json(df)
}
\arguments{
\item{df}{data.frame to represent as a JSON}
}
\value{
JSON string
}
\description{
Create a JSON representation of a data.frame
}
\details{
We represent a data.frame as an hash of hashes.
Factors are first coerced into characters.

The outer hash is keyed by the rownames of your data.frame
The inner hash is keyed by the colnames of your data.frame
}
\author{
Brad Friedman
}

