% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datanode.R
\name{data.frame.to.list}
\alias{data.frame.to.list}
\title{data.frame.to.json}
\usage{
data.frame.to.list(df)
}
\arguments{
\item{df}{data.frame to represent as a list}
}
\value{
list
}
\description{
Create a list representation of a data.frame
}
\details{
We represent a data.frame as an hash of hashes.
Factors are first coerced into characters.

The outer hash is keyed by the rownames of your data.frame
The inner hash is keyed by the colnames of your data.frame
}
\examples{
df <- data.frame(A=1:3, B=3:1, C=factor(c("foo","bar","foo")), row.names = c("one", "two", "three"))
## Should give the following
## list(one=list(A=1, B=3, C="foo"),
##      two=list(A=2, B=2, C="bar"),
##      three=list(A=3, B=1, C="foo"))
data.frame.to.list(df)
}
\author{
Brad Friedman
}

