% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param.set.R
\name{default.param}
\alias{default.param}
\title{default.param}
\usage{
default.param(name, prototype, ...)
}
\arguments{
\item{name}{Name of the parameter}

\item{prototype}{Default value on which the parameter should be built.}

\item{...}{Further arguments passed to the constructor for the appropriate parameter type. For example,
you can include \code{label}, \code{description}, \code{advanced} and \code{show.if}.}
}
\value{
AnalysisPageParam
}
\description{
Build a default AnalysisPageParam for one argument
}
\details{
You provide the name of the argument and default and I build an AnalysisPageParam.
The magic here is as follows:
\enumerate{
  \item{Named lists become \code{compound.param}, with \code{default.param()} then called recursively.}
  \item{Unnamed lists of become \code{array.param}, with \code{default.param()} called
        on the first element of the list to build the prototype. The length of the list
        is taken as the start value. min/max default to 0/Inf. (Advanced Note: the name is copied to the sub-element)}
  \item{Vectors of length > 1 become \code{select.param}s}
  \item{Vectors of length 0 or 1 and NULLs become \code{simple.param}s}
}
On any other type of argument it throws an error.
}
\examples{
default.param(name = "word", prototype = c("foo", "bar", "baz"), label = "Choose a word")
}
\author{
Brad Friedman
}

