% Generated by roxygen2 (4.0.2): do not edit by hand
\name{default.param.set}
\alias{default.param.set}
\title{default.param.set}
\usage{
default.param.set(handler)
}
\arguments{
\item{handler}{A function. (Typically one you are using as an AnalysisPage handler)}
}
\value{
AnalysisPageParamSet
}
\description{
Build a basic ParamSet for your handler
}
\details{
Each argument to your handler is rendered into a simple AnalysisPageParam with the name of the argument
and type "text". The idea is that you will then modify it as necessary to get more complicated widgets.
}
\examples{
f <- function(A=1, B=2) {}
  # param set with 2 form elements rendered as text inputs; something like   A [__________]     B [__________]
  pset <- default.param.set(f)
}
\author{
Brad Friedman
}

