% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service.R
\name{default.service.paramset}
\alias{default.service.paramset}
\title{default.service.paramset}
\usage{
default.service.paramset(handler)
}
\arguments{
\item{handler}{Handler function of service}
}
\value{
AnalysisPageParamSet
}
\description{
Create an AnalysisPageParamSet for a service handler
}
\details{
The services are not required to supply default values since they don't have to be rendered as parameters.
AnalysisPageServer still wants to have a param set. It can be used, for example, to check arguments
when building URLs. So we
artifically provide default values of 0 for everything, so you get a bunch of simple params.
}
\author{
Brad Friedman
}
