% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windows.R
\name{dieIfWindows}
\alias{dieIfWindows}
\title{Throw an error if platform is windows}
\usage{
dieIfWindows(caller = as.character(sys.call(-1))[1],
  errMesg = paste0(caller, "() does not run under windows"),
  isWindows = platformIsWindows())
}
\arguments{
\item{caller}{Name of calling function. Default: \code{as.character(sys.call(-1))[1]}}

\item{errMesg}{Error message to throw under windows. Default:
\code{paste0(caller, "() does not run under windows")}.}

\item{isWindows}{Boolean, indicating if current platform is windows.
Default: \code{platformIsWindows()}.}
}
\value{
Nothing
}
\description{
Throw an error if platform is windows
}

