% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dies.ok}
\alias{dies.ok}
\title{dies.ok}
\usage{
dies.ok(call, regex, testname)
}
\arguments{
\item{call}{An expression to evaluate}

\item{regex}{A regular expression to match the error against. If omitted then don't test the exception text.}

\item{testname}{A name for the test. Defaults to deparsing the call.}
}
\value{
Runs one or two tests (the second test to match the error message against regex, if it was provided and
if an error was successfully thrown).
}
\description{
Test that an expression throws an error
}
\details{
Test that an expression throws an error.
}
\examples{
dies.ok(stop("foo"), "foo", "it stops")
}
\author{
Brad Friedman
}

