% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/static.analysis.page.R
\name{embed.APS.dataset}
\alias{embed.APS.dataset}
\title{embed.APS.dataset}
\usage{
embed.APS.dataset(plot, df, title, show.sidebar = TRUE, show.table = TRUE,
  extra.html.classes = character(), extra.div.attr = character(),
  svg.args = list(), eval.args = list(envir = parent.frame()),
  outdir = get.APS.outdir(), randomize.filename = TRUE, ...)
}
\arguments{
\item{plot}{If present, then either an expression, a function, or a path to SVG file (not yet annotated).
If an then the expression will be evaluated after opening
a plotting device. The expression
will be evaluated in the calling frame, so your local variables will be
accessible, but this can be changed by modifying \code{eval.args}.
If a function, then the function will be called with no arguments. In that
case you would control the context yourself by setting the function's environment.
If path to an SVG then you would have already made the plot, and that would be used.
If missing then no plot is drawn---only the table is shown.}

\item{df}{data.frame of data. If omitted, then the return value
of evaluating the plotting expression or function is used (if \code{plot} is not a character).}

\item{title}{Caption for plot}

\item{show.sidebar}{Boolean, default TRUE. Set to FALSE to not show the
sidebar (filtering, tagging). (This is passed through directly to
\code{\link{aps.dataset.divs}}.)}

\item{show.table}{Boolean, default TRUE. Set to FALSE to not show the data
table (still available on download.
(This is passed through directly to
\code{\link{aps.dataset.divs}}.)}

\item{extra.html.classes}{Charvec of extra HTML classes to include in the
div. (This is embedded in a list then passed through directly to
\code{\link{aps.dataset.divs}}.)}

\item{extra.div.attr}{Names charvec of extra attributes to include in the div.
(This is embedded in a list then passed through directly to
\code{\link{aps.dataset.divs}}.)}

\item{svg.args}{Arguments (other than filename) to pass to the \code{svg}
function. This should be a named list. In particular, consider
something like \code{list(width = 8, height=5)} to change the aspect ratio.}

\item{eval.args}{Arguments to pass to \code{evalq} when evaluating your
plot code. Ignored if \code{plot} is character or a function. Otherwise
it should be a named list. Default is \code{list(envir = parent.frame())},
which means the evaluation will happen in the calling frame.}

\item{outdir}{Output directory. Default: \code{get.APS.outdir()}, which is
either "." or the directory of your knit2html target .html file.}

\item{randomize.filename}{Passed through to \code{\link{static.analysis.page}}
(but here the default is TRUE).}

\item{...}{Passed through to \code{static.analysis.page}.
t not \code{overwrite}, \code{outdir}, or \code{write.client}}
}
\value{
Returns the div, invisibly.
}
\description{
Embed an APS dataset
}
\details{
This function is meant to be called in a knitr document
that is being knit with \code{\link[knitr]{knit2html}}. It
makes a few assumptions that are valid in that context.

It makes a call to \code{\link{static.analysis.page}} for you
to annotate and write the SVG and JSON files, then emits
the \code{<div>} element to STDOUT. \code{outdir} defaults to ".".
It only does one plot/dataset at a time.
}
\examples{
message("See vignette embedding.html")
}
\author{
Brad Friedman
}

