% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datanode.R
\name{encode.datanode}
\alias{encode.datanode}
\title{JSON-Encode an AnalysisPageDataNode for the front end}
\usage{
encode.datanode(datanode)
}
\arguments{
\item{datanode}{AnalysisPageDataNode or other object}
}
\value{
JSON-encoded string
}
\description{
JSON-Encode an AnalysisPageDataNode for the front end.
This just calls \code{toJSON}, but before doing so
it makes sure that \code{$warnings} will be sent
as an array.
}
\details{
Mostly this function is only called once, from
the \code{$analysis} method of an \code{AnalysisPageRAapacheApp}.
}
\author{
Brad Friedman
}

