% Generated by roxygen2 (4.0.2): do not edit by hand
\name{event.names}
\alias{event.names}
\title{event.names}
\usage{
event.names(registry)
}
\arguments{
\item{registry}{EventRegistry}
}
\value{
Charvec
}
\description{
Get vector of names of all existing Events.
}
\details{
Get vector of names of all existing Events.
}
\examples{
r <- new.event.registry()
event.names(r)
add.event(r, "mouseclick")
event.names(r)
}
\author{
Brad Friedman
}

