% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static.analysis.page.R
\name{get.APS.libbase.prefix}
\alias{get.APS.libbase.prefix}
\title{Get current AnalysisPageServer library base directory}
\usage{
get.APS.libbase.prefix()
}
\value{
Path
}
\description{
Get current AnalysisPageServer library base directory. This is the
location that contains the JS, CSS, fonts, other files required
to render reports. The default "" means that these will always
be written in the directories containing individual reports and
datasets. Alternatively, if you are writing a lot of reports, you
can set this to a system-wide location (absolute path starting
and ending with "/") and then \code{\link{copy.front.end}} and
\code{setup.APS.knitr()} will use those instead.
}
\examples{
set.APS.libbase.prefix("/some/path")
get.APS.libbase.prefix()
}
\seealso{
\code{\link{set.APS.libbase.prefix}}
}
\author{
Brad Friedman
}
