% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{getTraceback}
\alias{getTraceback}
\title{getTraceback}
\usage{
getTraceback(mto)
}
\arguments{
\item{mto}{An object of the try-error class}
}
\value{
Traceback as a string
}
\description{
Get traceback from tryKeepTraceback()
}
\examples{
x <- tryKeepTraceback(stop("no way"))
if(is(x, "try-error"))  cat(getTraceback(x))
}

