% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event.registry.R
\name{has.event}
\alias{has.event}
\title{has.event}
\usage{
has.event(registry, event)
}
\arguments{
\item{registry}{EventRegistry}

\item{event}{String. Name of the Event.}
}
\value{
Logical
}
\description{
Predicate to test if an EventRegistry has an Event of a given name
}
\details{
Predicate to test if an EventRegistry has an Event of a given name
}
\examples{
r <- new.event.registry()
has.event(r, "mouseclick")
add.event(r, "mouseclick")
has.event(r, "mouseclick")
}
\author{
Brad Friedman
}
