% Generated by roxygen2 (4.0.2): do not edit by hand
\name{has.page}
\alias{has.page}
\alias{has.page.AnalysisPageRegistry}
\title{has.page}
\usage{
has.page(registry, page.name)

\method{has.page}{AnalysisPageRegistry}(registry, page.name)
}
\arguments{
\item{registry}{Registry}

\item{page.name}{An AnalysisPageRegistry}
}
\value{
Logical, indicating if a page is already registered under than name
}
\description{
Predicate to test if some page is already registered under a name
}
\details{
Predicate to test if some page is already registered under a name
}
\examples{
example(register.page, ask=FALSE)  # register the sine page
has.page(registry, "sine")         # should return TRUE now.
}
\author{
Brad Friedman
}
\seealso{
\code{\link{new.registry}}, \code{\link{register.page}}, \code{\link{get.page}}, \code{\link{pages}}
}

