% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test.R
\name{ignore.lots.of.stuff}
\alias{ignore.lots.of.stuff}
\title{ignore.lots.of.stuff}
\usage{
ignore.lots.of.stuff(lines)
}
\arguments{
\item{lines}{Character vector of lines from the SVG file.}
}
\value{
Character vector. Same lines, with id and class tags transformed.
}
\description{
Transformer for ignoring id, class, type and some whitespace
}
\details{
This transformer strips all id, class and type tags, with one preceding space, from the SVG lines.

It also ignores what it thinks is space between tags, namely >\\s+<

This is meant primary as an argument for \code{transformer} in \code{\link{check.same.svgs}}.

Not exported---you should fully qualify it with \code{AnalysisPageServer:::ignore.lots.of.stuff} if you
want to use it.

All the lines will be concatenated, too, into a single character string.
}
\author{
Brad Friedman
}
\seealso{
\code{\link{check.same.svgs}}
}

