% Generated by roxygen2 (4.0.2): do not edit by hand
\name{is.registry}
\alias{is.registry}
\title{is.registry}
\usage{
is.registry(registry)
}
\arguments{
\item{registry}{A candidate object}
}
\value{
Logical, indicating that the object is an "AnalysisPageRegistry"
}
\description{
Test if an argument is an AnalysisPageRegistry
}

