% Generated by roxygen2 (4.0.2): do not edit by hand
\name{kill.process}
\alias{kill.process}
\title{kill.process}
\usage{
kill.process(pid)
}
\arguments{
\item{pid}{Process ID, or list with \code{$pid} component}
}
\value{
Same as \code{\link[fork]{wait}}.
}
\description{
Kill a process and wait for it.
}
\details{
Kill a process and wait for it. Nothing more than \code{kill(pid); wait(pid)},
but handy to have a single function so you don't forget the \code{wait()} call.
}
\author{
Brad Friedman
}

