% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr.R
\name{knitAPS}
\alias{knitAPS}
\title{Render an AnalysisPageServer Rmd report to HTML using knitr}
\usage{
knitAPS(template.Rmd, outdir = dirname(outfile), outfile = file.path(outdir,
  sub(".Rmd", ".html", basename(template.Rmd))), e = new.env(),
  overwrite = FALSE, ...)
}
\arguments{
\item{template.Rmd}{Path to .Rmd template}

\item{outdir}{Path to output directory. Default: \code{dirname(outfile)}. You must
specify either this or \code{outfile}.}

\item{outfile}{Path to output main .html file. Default is to use \code{basename(template.Rmd)}
with an .html extension, and put it in the \code{outdir} directory.
You must specify either \code{outdir} or this.}

\item{e}{Environment to pass to knitr. Default: \code{new.env()}}

\item{overwrite}{Logical, default FALSE. This controls the behavior If outdir already exists.
In that case, if \code{overwrite} is TRUE then the outdir is first removed. If \code{overwrite}
is FALSE then instead an error is thrown.}

\item{...}{Passed through to \code{\link[knitr]{knit2html}()}}
}
\value{
Path to filename written (that is, \code{outfile}).
}
\description{
This doesn't do much more than just call knitr, but it first
has a bit of logic to overwrite or not the target directory. AnalysisPageServer
reports are entire directories, so this is convenient to manage the creation
of the report at that level.
}
\author{
Brad Friedman
}
