% Generated by roxygen2 (4.0.2): do not edit by hand
\name{known.param.sizes}
\alias{known.param.sizes}
\title{known.param.sizes}
\usage{
known.param.sizes()
}
\value{
Character vector
}
\description{
Get the controlled vocabulary of parameter size words
}
\details{
Returns the controlled vocabulary of parameter size words
}
\author{
Brad Friedman
}

