% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test.R
\name{lives.ok}
\alias{lives.ok}
\title{lives.ok}
\usage{
lives.ok(call, testname)
}
\arguments{
\item{call}{An expression to evaluate}

\item{testname}{A name for the test. Defaults to deparsing the call.}
}
\value{
Runs one test. Returns the value of the evaluated expression
}
\description{
Test that an expression lives OK
}
\details{
Test that evaluating an expression lives OK (does not throw an exception)
}
\examples{
lives.ok(3+5, "addition lives OK")
}
\author{
Brad Friedman
}

