% Generated by roxygen2 (4.0.2): do not edit by hand
\name{new.datanode.array}
\alias{new.datanode.array}
\title{new.datanode.array}
\usage{
new.datanode.array(name, children, ...)
}
\arguments{
\item{name}{Name of the node}

\item{children}{List of \code{AnalysisPageDataNode}s}

\item{...}{Passed through to \code{new.datanode}, in particular \code{label} and \code{description}}
}
\value{
AnalysisPageDataNode
}
\description{
Construct a new array-type data node
}
\details{
Construct a new array-type data node from a list of \code{AnalysisPageDataNode}s
}
\examples{
html.node <- example(new.datanode.html)$value
simple.node <- example(new.datanode.simple)$value
new.datanode.array(name = "arr", children = list(html.node, simple.node))
}
\author{
Brad Friedman
}

