% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datanode.R
\name{new.datanode.html}
\alias{new.datanode.html}
\title{new.datanode.html}
\usage{
new.datanode.html(name, value, ...)
}
\arguments{
\item{name}{Name of the node}

\item{value}{The value, an HTML string or charvec.}

\item{...}{Passed through to \code{new.datanode}, in particular \code{label} and \code{description}}
}
\value{
AnalysisPageDataNode
}
\description{
Construct a new HTML-type data node
}
\details{
Construct a new HTML-type data node. An HTML-type data node is like a simple data node but
it has an HTML string or character vector as its value. It should be rendered as-is, but
with activated analysis-page-data-set containers.
}
\examples{
new.datanode.html(name = "shakespeare",
                  value = "<i>Shall I compute thee to a summer's data?</i>")
}
\author{
Brad Friedman
}
\seealso{
\code{\link{is.atomic}}
}

