% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datanode.R
\name{new.datanode.plot}
\alias{new.datanode.plot}
\title{new.datanode.plot}
\usage{
new.datanode.plot(name, plot.file, table, ...)
}
\arguments{
\item{name}{Name of the node}

\item{plot.file}{Path to plot file, relative to server tempdir.}

\item{table}{A table-type \code{AnalysisPageDataNode}}

\item{...}{Passed through to \code{new.datanode}, in particular \code{label} and \code{description}}
}
\value{
AnalysisPageDataNode
}
\description{
Construct a new plot-type data node
}
\details{
Construct a new plot-type data node from a plot.file and an AnnotatedDataFrame.
Note: caption is included implicitly in the table object.
}
\author{
Brad Friedman
}

