% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datanode.R
\name{new.datanode.simple}
\alias{new.datanode.simple}
\title{new.datanode.simple}
\usage{
new.datanode.simple(name, value, ...)
}
\arguments{
\item{name}{Name of the node}

\item{value}{The value}

\item{...}{Passed through to \code{new.datanode}, in particular \code{label} and \code{description}}
}
\value{
AnalysisPageDataNode
}
\description{
Construct a new simple-type data node
}
\details{
Construct a new simple-type data node. A simple data node must have either a NULL value
or a "scalar" value, which means length 1 atomic
}
\examples{
new.datanode.simple(name = "x", value = 100)
}
\author{
Brad Friedman
}
\seealso{
\code{\link{is.atomic}}
}

