% Generated by roxygen2 (4.0.2): do not edit by hand
\name{new.datanode.table}
\alias{new.datanode.table}
\title{new.datanode.table}
\usage{
new.datanode.table(name, data, caption = "", ...)
}
\arguments{
\item{name}{Name of the node}

\item{data}{An \code{AnnotatedDataFrame}. Unless "label" is already available,
"labelDescription" in the \code{varMetadata} will be
changed to "label" to agree with the syntax we use in the rest of the AnalysisPage interface.}

\item{caption}{Caption for the table. Default: ""}

\item{...}{Passed through to \code{new.datanode}, in particular \code{label} and \code{description}}
}
\value{
AnalysisPageDataNode
}
\description{
Construct a new table-type data node
}
\details{
Construct a new table-type data node either from an \code{AnnotatedDataFrame}.
}
\note{
Captions for plots are included implicitly in the table component of their responses.
}
\author{
Brad Friedman
}

