% Generated by roxygen2 (4.0.2): do not edit by hand
\name{new.event.registry}
\alias{new.event.registry}
\title{new.event.registry}
\usage{
new.event.registry()
}
\value{
EventRegistry
}
\description{
EventRegistry constructor
}
\details{
An EventRegistry is a collection of Events. Each Event has a name
and a list of functions, each known as a handler.
Events can be modified or triggered. When triggered, each function
is called in turn, and the final return value is returned to the triggering context.
}
\examples{
r <- new.event.registry()
}
\author{
Brad Friedman
}

