% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/registry.R
\name{new.registry}
\alias{new.registry}
\title{new.registry}
\usage{
new.registry(...)
}
\arguments{
\item{...}{\code{AnalysisPage}s with which to initially populate the registry}
}
\value{
A new registry, which is just an empty list with class AnalysisPageRegistry.
}
\description{
Make a new (empty) registry.
}
\examples{
r <- new.registry()
}
\author{
Brad Friedman
}
\seealso{
\code{\link{register.page}}, \code{\link{has.page}}, \code{\link{get.page}}, \code{\link{pages}}
}

