% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{new.response}
\alias{new.response}
\title{new.response}
\usage{
new.response(body, content.type, status = 200, headers = character())
}
\arguments{
\item{body}{Either a raw vector or a character vector that constitutes the response body.}

\item{content.type}{A string giving the content-type, such as "text/plain"}

\item{status}{Integer. An HTTP response status. Default, 200, means HTTP_OK}

\item{headers}{Named charvec of extra HTTP headers. Default: \code{character()}  (none)}
}
\value{
AnalysisPageResponse object
}
\description{
Build a new AnalysisPageResponse object
}
\details{
A handler may return an AnalysisPageResponse object, which is basically a complete
response, if it doesn't want the framework to do any extra processing. This allows
complete control over the response.
}
\examples{
poem.file <- system.file("examples/in-a-station-of-the-metro.html", package="AnalysisPageServer")
poem.html <- readLines(poem.file, warn = FALSE)
new.response(paste0(poem.html, "\\n"), content.type = "text/html")
}
\author{
Brad Friedman
}
