% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rook.analysis.page.server.R
\name{new.rook.analysis.page.app}
\alias{new.rook.analysis.page.app}
\title{new.rook.analysis.page.app}
\usage{
new.rook.analysis.page.app(app, EP = NULL, front.end.location = "/dist-aps",
  front.end.dir = system.file("htdocs/client", package =
  "AnalysisPageServer"), app.name = "RAPS",
  app.prefix = file.path("/custom", app.name), tmpdir = tempdir(), ...)
}
\arguments{
\item{app}{AnalysisPageRApacheApp. Or an AnalysisPageRegistry from which to build an app}

\item{EP}{ExpressionPlotClient object, if needed for your app.}

\item{front.end.location}{Location relative to App base directory from which to serve
front end files. Default: "/dist-aps".}

\item{front.end.dir}{Path (in filesystem) to front end files. Default:
\code{system.file("htdocs", package = 'AnalysisPageServer')}}

\item{app.name}{The name of the app you are going to use within Rook. This is used to build the prefix \code{/custom/$\{app.name\}}
from where the app will be served---the Javascript front end has to be notified of this. Ignored if \code{app.prefix} is supplied.}

\item{app.prefix}{The prefix from which the app will be served. Default: \code{/custom/$\{app.name\}}.}

\item{tmpdir}{Path to temporary directory to store files needed while the server is running. Default: \code{tempdir()}.}

\item{...}{If \code{app} is actually an AnalysisPageRegistry then \code{...} is passed through
along with it to \code{rapache.app.from.registry} to build the \code{AnalysisPageRApacheApp}.}
}
\value{
Your app, as a Rook App
}
\description{
Given an RApacheAnalysisPageServer object, return a Rook app that can run it.
}
\details{
new.rook.analysis.page.app
}
\examples{
message("See vignette ExamplesServers.html")
}
\seealso{
\code{\link{startRookAnalysisPageServer}}, \code{\link{kill.process}}
}
\author{
Brad Friedman
}
