% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry-S3generics.R, R/registry.R
\name{pages}
\alias{pages}
\alias{pages.AnalysisPageRegistry}
\title{pages}
\usage{
pages(registry, include.services = FALSE)

\method{pages}{AnalysisPageRegistry}(registry, include.services = FALSE)
}
\arguments{
\item{registry}{AnalysisPageRegistry object}

\item{include.services}{Logical. Should I include services in my list of all pages? Default: FALSE, do not include services.}
}
\value{
Character vector of names of pages in registry
}
\description{
Get names of all pages in registry
}
\details{
Get names of all pages in registry
}
\note{
Service pages are identified as those having their \code{service} flag set, which is done at page build time using
the \code{service} parameter of the \code{new.analysis.page} constructor.
}
\examples{
empty.pages <- pages(new.registry())   # should be empty character vector
example(register.page, ask=FALSE)      # see register.page example---registers the sine handler
pages(registry)                        # should now be the character vector "sine"
}
\author{
Brad Friedman
}
\seealso{
\code{\link{new.registry}}, \code{\link{register.page}}, \code{\link{has.page}}, \code{\link{get.page}}
}

