% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/param.set.R
\name{param.set}
\alias{param.set}
\title{param.set}
\usage{
param.set(...)
}
\arguments{
\item{...}{AnalysisPageParam objects. Or a single argument, being a list of AnalysisPageParam objects.}
}
\value{
AnalysisPageParamSet
}
\description{
Build a new AnalysisPageParamSet from a list of AnalysisPageParam's.
}
\details{
Build a new AnalysisPageParamSet from a list of AnalysisPageParam's: Check for non-AnalysisPageParam elements
and for duplicate names. Apply $name elements to names of param set. Set class to AnalysisPageParamSet
}
\examples{
par1 <- simple.param(name = "par1")
par2 <- bool.param(name = "par2")
pset <- param.set(par1, par2)
}
\author{
Brad Friedman
}

