% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/param.set.R
\name{paramSetToJSON}
\alias{paramSetToJSON}
\title{Convert an AnalysisPageParamSet to a JSON string}
\usage{
paramSetToJSON(ps)
}
\arguments{
\item{ap}{AnalysisPageParameter}
}
\value{
JSON string
}
\description{
This is almost just calling toJSON but it knows to first remove $transformer
components, since functions can't be JSON encoded, and anyway that
is really server-side information.
}
\author{
Brad Friedman
}

