% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistent.params.R
\name{persistent.param.dependencies}
\alias{persistent.param.dependencies}
\alias{persistent.param.dependencies.AnalysisPage}
\alias{persistent.param.dependencies.AnalysisPageParam}
\alias{persistent.param.dependencies.AnalysisPageParamSet}
\alias{persistent.param.dependencies.AnalysisPageRegistry}
\title{Return persistent parameter dependencies for an object}
\usage{
persistent.param.dependencies(x)

\method{persistent.param.dependencies}{AnalysisPageParam}(x)

\method{persistent.param.dependencies}{AnalysisPageParamSet}(x)

\method{persistent.param.dependencies}{AnalysisPage}(x)

\method{persistent.param.dependencies}{AnalysisPageRegistry}(x)
}
\arguments{
\item{x}{An object for which there is a \code{persistent.param.dependencies} method.}
}
\value{
Named list of charvecs, see Details.
}
\description{
Return persistent parameters dependencies for an object.

For an \code{AnalysisPageParam} the return value is either
a charvec of the object's
persistent dependencies. It is important to note that
The namespace for persistent dependencies
of an \code{AnalysisPageParam} is the same as the namespace
of the page Parameters. This is because when calling a constructor
for an \code{AnalysisPageParam} at most we might specify the
names of other Parameters (for example for combobox dependencies).
But the other Parameters are not available at that time, so
we can't look up their corresponding names in the persistent
namespace.

For an \code{AnalysisPageParamSet} or an
\code{AnalysisPage}, the return value is a list
named after all of the persistent params, the values being
charvecs of their dependencies. Note that
The namespace for persistent dependencies
of an \code{AnalysisPageParamSet} is the persistent namespace,
not the Page Parameter namespace.

For an \code{AnalysisPageRegistry}
the return value is a list
named after all of the persistent params of any of the pages,
the values being
charvecs of their dependencies. If there is an discrepancy in
the dependencies for a given persistent param, then the union of
all dependencies is taken. (This will be checked for acyclicity
by \code{.validate.registry}.)
The namespace for persistent dependencies
of an \code{AnalysisPageRegistry} is the persistent namespace,
not the Page Parameter namespace.
}
\details{
This is a named list. The names are the names of persistent
parameters, and the values are the other persistent parameters
on which they depend.
}
\author{
Brad Friedman
}

