% Generated by roxygen2 (4.0.2): do not edit by hand
\name{persistent.params}
\alias{persistent.params}
\alias{persistent.params.AnalysisPage}
\alias{persistent.params.AnalysisPageParam}
\alias{persistent.params.AnalysisPageParamSet}
\alias{persistent.params.AnalysisPageRegistry}
\title{Return names of persistent parameters for an object}
\usage{
persistent.params(x)

\method{persistent.params}{AnalysisPageParam}(x)

\method{persistent.params}{AnalysisPageParamSet}(x)

\method{persistent.params}{AnalysisPage}(x)

\method{persistent.params}{AnalysisPageRegistry}(x)
}
\arguments{
\item{x}{An object for which there is a \code{persistent.params} method.}
}
\value{
Charvec of persistent params
}
\description{
Return names of persistent parameters for an object. The exact
meaning depends on the type of object. For an AnalysisPage it would
be the persistent params for that page. For an AnalysisPageRegistry
it would be the persistent params for any of its pages.
}
\details{
Note that the names are from the shared namespace, which are
the values passed as the \code{persistent} argument to the
\code{AnalysisPageParam} constructor functions such as \code{\link{simple.param}}.
Although this is usually the same as actual names of the \code{AnalysisPageParam}s themselves
there is no requirement that they be the same.
}
\author{
Brad Friedman
}

