% Generated by roxygen2 (4.0.2): do not edit by hand
\name{platformIsWindows}
\alias{platformIsWindows}
\title{Predicate to test if running on a windows platform}
\usage{
platformIsWindows()
}
\value{
Boolean:  \code{TRUE} on windows, \code{FALSE} on other platforms
}
\description{
Predicate to test if running on a windows platform
}
\author{
Brad Friedman
}

