% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rapache.R
\name{rapache.trig.app}
\alias{rapache.trig.app}
\title{rapache.trig.app}
\usage{
rapache.trig.app(...)
}
\arguments{
\item{...}{Other parameters to pass through to \code{rapache.app.from.registry}, such as \code{con}}
}
\value{
AnalysisPageRegistry
}
\description{
Build the AnalysisPageRApacheApp for the trig example
}
\details{
The toy registry has a sine page, a cosine page and the scattergram tool.
}
\author{
Brad Friedman
}
\seealso{
\code{\link{trig.registry}}
}

