% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/event.registry.R
\name{remove.event}
\alias{remove.event}
\title{remove.event}
\usage{
remove.event(registry, event)
}
\arguments{
\item{registry}{EventRegistry}

\item{event}{String. Name of the Event to remove}
}
\value{
Nothing good.
}
\description{
Remove an Event entirely
}
\details{
Remove an Event entirely from the EventRegistry. Contrast with \code{\link{clear.event.handlers}},
which only removes the handlers for that event.
}
\examples{
r <- new.event.registry()
add.event(r, "mouseclick")
has.event(r, "mouseclick")
remove.event(r, "mouseclick")
has.event(r, "mouseclick")
}
\author{
Brad Friedman
}

