% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapache.R
\name{remove.old.files}
\alias{remove.old.files}
\title{remove.old.files}
\usage{
remove.old.files(tmpdir, tmpdir.timeout.seconds)
}
\arguments{
\item{tmpdir}{Path to directory whose old files you want to delete}

\item{tmpdir.timeout.seconds}{Time in seconds. An attempt will be made to
delete files with ctimes older than this many seconds before the current time.}
}
\value{
see \code{\link{unlink}}
}
\description{
Remove old files from a directory
}
\details{
Remove old files from a directory
}
\author{
Brad Friedman
}
