% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.page.R
\name{getCustomContent}
\alias{appendCustomContent}
\alias{clearRequestEnv}
\alias{getCustomContent}
\title{Functions to manage custom content other aspects of the request-specific environment}
\usage{
getCustomContent()

appendCustomContent(sectionName, content)

clearRequestEnv()
}
\arguments{
\item{sectionName}{Name of section (string)}

\item{content}{Character vector of HTML content to append}
}
\value{
\code{getCustomContent} returns named list of character vectors

\code{appendCustomContent} does not return anything good.

\code{clearRequestEnv} does not return anything useful
}
\description{
Custom content are HTML rendered
as additional accordion sections. From the data structure
point of view these are represented as a named of list
of character vectors. The names are the section headers.
Use \code{\link{appendCustomContent}} to add more content.

\code{appendCustomContent} adds custom content to be rendered in separate accordion section

\code{clearRequestEnv} clears the environment associated with the last request.
}
\examples{
appendCustomContent(sectionName = "foo", content = c("<i>bar</i><br>","<b>baz</b>"))
getCustomContent()
clearRequestEnv()
}
\author{
Brad Friedman
}

