% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/static.analysis.page.R
\name{reset.APS.outdir}
\alias{reset.APS.outdir}
\title{Reset AnalysisPageServer output directory}
\usage{
reset.APS.outdir()
}
\value{
Nothing of note
}
\description{
This directory is used by \code{\link{embed.APS.dataset}} to decide
where to save the .svg and .json files. This function resets it to
its default, ".".
}
\examples{
set.APS.outdir("/some/path")
get.APS.outdir()
reset.APS.outdir()
}
\author{
Brad Friedman
}
\seealso{
\code{\link{get.APS.outdir}}, \code{\link{set.APS.outdir}}
}

