% Generated by roxygen2 (4.0.2): do not edit by hand
\name{select.param}
\alias{select.param}
\title{select.param}
\usage{
select.param(..., value, choices, allow.multiple = FALSE,
  style = "dropdown")
}
\arguments{
\item{...}{Passed through to \code{\link{simple.param}}. This includes at least "name", optionally "label" and "description"
t not "type".}

\item{value}{Default value. If not specified then the first entry in choices is taken to be the default.}

\item{choices}{A character vector giving the choices to display. If named, then the values are used for display and names are used
for the actual form values. If unnamed, then the values are used for both display and names.}

\item{allow.multiple}{If TRUE then render as checkbox group and allow multiple selections. (The function
will be provided a vector of all selected values.) If \code{allow.multiple} is TRUE then \code{style} must be "dropdown".
Default: FALSE}

\item{style}{Either "dropdown" (default), to render as dropdown list, or "radio", to render as radio group.
If \code{allow.multiple} is TRUE then \code{style} must be "dropdown".}
}
\value{
An AnalysisPageParam
}
\description{
Build a select AnalysisPageParam
}
\details{
Build a select AnalysisPageParam. This is probably rendered either as a dropdown or radio group. It is a selection from a fixed
list of possible values. The list is known before page load time
}
\examples{
color <- select.param("color", label="Color", description="Color of your house", choices=c("red","green","mauve","tope"))
}
\author{
Brad Friedman
}

