% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.R
\name{service.link}
\alias{service.link}
\title{Build a URL to call a AnalysisPageServer webservice}
\usage{
service.link(page, params = list(), app.base.url)
}
\arguments{
\item{page}{Name of page}

\item{params}{List of parameter values (as R objects---this function will encode them). Default: \code{list()} (no parameters).}

\item{app.base.url}{Base URL for application. This is usually the prefix in which the app landing HTML page is found.}
}
\value{
URL
}
\description{
This function is simply a specialization of \code{analysis.link} with a few conveniences
for webservice-type pages. In particular, the parameters of that function about plotting
are not available.
}
\author{
Brad Friedman
}

