% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static.analysis.page.R
\name{set.APS.libbase.prefix}
\alias{set.APS.libbase.prefix}
\title{Set current AnalysisPageServer library base directory}
\usage{
set.APS.libbase.prefix(libbase.prefix)
}
\arguments{
\item{libbase.prefix}{New libbase.prefix. Must either be empty string
or end with "/"}
}
\value{
libbase.prefix, again
}
\description{
Set current AnalysisPageServer library base directory.
See \code{\link{get.APS.libbase.prefix}()} for more information.
}
\examples{
set.APS.libbase.prefix("/some/path/")
get.APS.libbase.prefix()
}
\seealso{
\code{\link{get.APS.libbase.prefix}}
}
\author{
Brad Friedman
}
