% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static.analysis.page.R
\name{set.APS.outdir}
\alias{set.APS.outdir}
\title{Set current AnalysisPageServer output directory}
\usage{
set.APS.outdir(outdir)
}
\arguments{
\item{outdir}{New output directory}
}
\value{
Nothing important
}
\description{
This directory is used by \code{\link{embed.APS.dataset}} to decide
where to save the .svg and .json files.
}
\note{
It seems like it would be a good idea to follow this call with an \code{on.exit(reset.APS.outdir())}.
But \code{on.exit} within a knitr chunk it will just first at the end of the chunk. If you are using
knitr then you should just call \code{setup.APS.knitr()} at the top of your document then each document
will have its output directory correctly set and you don't really have to worry. If you want to be
really anal you could call \code{reset.APS.outdir()} at the bottom of your knitr document.
}
\examples{
set.APS.outdir("/some/path")
get.APS.outdir()
reset.APS.outdir()
}
\seealso{
\code{\link{get.APS.outdir}}, \code{\link{reset.APS.outdir}}
}
\author{
Brad Friedman
}
