% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.page.R
\name{sine.handler}
\alias{sine.handler}
\title{sine.handler}
\usage{
sine.handler(xmin = 0, xmax = 3 * pi, n = 100)
}
\arguments{
\item{xmin}{Numeric. Minimum x value to plot}

\item{xmax}{Numeric. Maximum x value to plot}

\item{n}{Integer. Number of points to plot}
}
\value{
data.frame
}
\description{
An example handler just for testing and development
}
\details{
This handler takes three parameters, xmin, xmax and n,
makes a plot of the sin curve from xmin to xmax (using
n equally spaced points), and returns a data.frame
with the x and y coordinates, with IDs A-Z, A.1-Z.1, ...
}
\author{
Brad Friedman
}

