% Generated by roxygen2 (4.0.2): do not edit by hand
\name{startRookAnalysisPageServer}
\alias{startRookAnalysisPageServer}
\title{Start a new Rook AnalysisPage server}
\usage{
startRookAnalysisPageServer(reg, tmpdir = tempdir(), ...,
  app = new.rook.analysis.page.app(reg, tmpdir = tmpdir, app.name = app.name,
  ...), app.name = "RAPS", port = 5000)
}
\arguments{
\item{reg}{AnalysisPageRegistry from which to build application. Passed through to \code{\link{new.rook.analysis.page.app}}.}

\item{tmpdir}{Directory for temporary files. Passed through to \code{\link{new.rook.analysis.page.app}}. Default: \code{tempdir()}.}

\item{...}{Passed through to \code{\link{new.rook.analysis.page.app}}.}

\item{app}{Rook App to put into the server. Default: \code{new.rook.analysis.page.app(reg, tmpdir = tmpdir, ...)}.
Normally you would omit this argument.}

\item{app.name}{Name for App within server, default "RAPS" (for Rook AnalysisPageServer).
This will determine the second part of the URL, for example "/custom/RAPS".}

\item{port}{Port on which to start listening.}
}
\value{
list with two componenets:
\describe{
  \item{\code{$url}}{URL to base of application}
  \item{\code{$pid}}{Process ID of server}
}
}
\description{
Start a new Rook AnalysisPage server. This is a convenience
wrapper around \code{\link{new.rook.analysis.page.app}}
which builds the Rook App and then also makes a Rook server (\code{Rhttpd} object)
which just contains the one App. It then starts the server in a fork and returns the PID
of the child process.
}
\note{
This function used to be called \code{start.rook.analysis.page.server} but that
led to an R CMD check warning about S3 method inconsistency.
}
\examples{
\dontrun{
  registry <- AnalysisPageServer:::trig.registry()
  server <- startRookAnalysisPageServer(registry, port = 5102)

  ## do some stuff
  ## For example
  landing.page.url <- rook.analysis.page.server.landing.page(server)
  ## now go to your web browser and open landing.page.url

  ## Or maybe something in this R process. See what the pages are
  pages.url <- file.path(server$url, "R", "pages")
  pages <- fromJSON(readLines(pages.url, warn = FALSE))
  sapply(pages, "[[", "name")

  ## Kill the server
  kill.process(server)
}
message("See vignette ExamplesServers.html")
}
\author{
Brad Friedman
}
\seealso{
\code{\link{new.rook.analysis.page.app}}, \code{\link{kill.process}}
}

