% Generated by roxygen2 (4.0.2): do not edit by hand
\name{static.analysis.page.query.string}
\alias{static.analysis.page.query.string}
\title{static.analysis.page.query.string}
\usage{
static.analysis.page.query.string(paths.list)
}
\arguments{
\item{paths.list}{\code{paths.list} is (for example) the return value from
\code{.write.plots.and.data.for.static.analysis.page}. It is a list whose entries
correspond to the datasets on your page. Each entry is in turn a list with a \code{$plot}
and/or \code{$data} element, each of which is a URL (but could be relative
to the application \code{.html} file) to the encoded SVG and JSON data files.}
}
\value{
Query string, starting with "#"
}
\description{
Build the query string for a static analysis page
}
\details{
All static analysis pages are deployed on top of the same HTML/Javascript/CSS stack.
To point the client to the correct plots and data, their paths are encoded into the
query part of the URL. This function performs that encoding.

The query string will begin with "#".

To form a URL to view your data, simply append this query string to the URL for the application
\code{.html} file.
}
\author{
Brad Friedman
}

