% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test.package}
\alias{test.package}
\title{test.package}
\usage{
test.package(pattern = "^test.*R$", package = "AnalysisPageServer")
}
\arguments{
\item{pattern}{String. Regular expression. Only filenames matching}

\item{package}{Name of package to test. Default: "AnalysisPageServer"
this expression will be included in the test harness. Default:
\code{"^test.*R$"}.}
}
\value{
RUnitTestData
}
\description{
Run the RUnit test harness for this package
}
\details{
Run the RUnit test harness for this package
}
\author{
Brad Friedman, Cory Barr
}
\seealso{
\code{\link[RUnit]{runTestSuite}}, \code{\link[base]{require}}
}

