% Generated by roxygen2 (4.0.2): do not edit by hand
\name{trigger.event}
\alias{trigger.event}
\title{trigger.event}
\usage{
trigger.event(registry, event, ...)
}
\arguments{
\item{registry}{EventRegistry}

\item{event}{Name of event to trigger}

\item{...}{Further parameters are passed to each handler in turn.}
}
\value{
See Details
}
\description{
Trigger a registered Event
}
\details{
Trigger a registered Event.

Every handler is called in turn. If any handler returns a value with a "CatchEvent"
attribute set to TRUE then no further handlers are called. That attribute is removed
from the return value and the value is returned to the triggering context. Otherwise
the return value of only the last function is called.

If no handlers are registered then NULL is returned.

If no Event exists of that name then an error is thrown.
}
\examples{
r <- new.event.registry()
add.event(r, "mouseclick")
add.event.handler(r, "mouseclick", function(x, y)  message("Mouse clicked at coordinates (", x, ", ", y, ")"))
trigger.event(r, "mouseclick", x = 30, y = 50)
}
\author{
Brad Friedman
}

