% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/error.R
\name{tryKeepConditions}
\alias{AnalysisPageValueWithConditions}
\alias{tryKeepConditions}
\title{tryKeepConditions}
\usage{
tryKeepConditions(expr)
}
\arguments{
\item{expr}{Expression to evaluate}
}
\value{
AnalysisPageValueWithConditions
}
\description{
Try-catch wrapper, keeping error traceback and conditions
}
\details{
This is a try-catch wrapper. It returns a list with
four elements:
\describe{
  \item{\code{$value}}{The value of the evaluated expression, or NULL if an error was thrown and execution did not complete}
  \item{\code{$messages}}{A list of \code{message} objects, each being a message thrown during the execution, in order}
  \item{\code{$warnings}}{A list of \code{condition} objects, each being a message thrown during the execution, in order}
  \item{\code{$error}}{NULL if there was no error, otherwise the \code{error} object, which can then be passed to \code{getTraceback}
        to retrieve the error}
}

The elements of the \code{$messages}, \code{$warnings} and \code{$error} are all actually two-element lists, the first being
the condition object itself (named \code{$message}, \code{$warning} or \code{$error}) and the second begin the call stack
as returned by \code{sys.calls()} and named \code{$calls}.

The class of this object is set as "AnalysisPageValueWithConditions"
}
\examples{
value.with.warning <- tryKeepConditions({warning("warning message"); 3})
value.with.error <- tryKeepConditions({stop("err message")})
}
\author{
Brad Friedman
}
\seealso{
\code{vwc.is.error}

\code{\link{vwc.conditions}} \code{\link{vwc.error}} \code{\link{vwc.error.condition}} \code{\link{vwc.error.traceback}} \code{\link{vwc.is.error}} \code{\link{vwc.messages}} \code{\link{vwc.messages.conditions}} \code{\link{vwc.messages.tracebacks}} \code{\link{vwc.n}} \code{\link{vwc.n.messages}} \code{\link{vwc.n.warnings}} \code{\link{vwc.tracebacks}} \code{\link{vwc.value}} \code{\link{vwc.warnings}} \code{\link{vwc.warnings.conditions}} \code{\link{vwc.warnings.tracebacks}}
}

