% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/error.R
\name{tryKeepTraceback}
\alias{tryKeepTraceback}
\title{tryKeepTraceback}
\usage{
tryKeepTraceback(expr)
}
\arguments{
\item{expr}{Expression to evaluate}
}
\value{
Result of expression or error if thrown
}
\description{
Wrapper around try-catch
}
\examples{
x <- tryKeepTraceback(stop("no way"))
if(is(x, "try-error"))  cat(getTraceback(x))
}

