% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/provide.urlEncode.R
\name{urlEncode}
\alias{urlEncode}
\title{urlEncode a string}
\arguments{
\item{x}{Character vector of strings to urlEncode}
}
\value{
Character vector of same length as \code{x} containing
encoded strings
}
\description{
When a function of this name is available from the global environment
(such as when running under RApache) then that function is used.
Otherwise a pure R implementation is provided.
}
\author{
Brad Friedman
}

