% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/validate.param.list.R
\name{validate.array.param.value}
\alias{validate.array.param.value}
\title{validate.compound.param.value}
\usage{
validate.array.param.value(app, val, transform.labeled = FALSE)
}
\arguments{
\item{app}{AnalysisPageParam}

\item{val}{Candidate value}

\item{transform.labeled}{Passed through to \code{\link{validate.param.value}}. Default: \code{FALSE}.}
}
\value{
value, possibly with elements transformed
}
\description{
Validate an array-type AnalysisPageParam value
}
\details{
Validate an array-type AnalysisPageParam value:
\enumerate{
  \item \code{val} must be a list.
  \item \code{length(val)} must be in the acceptable range (between \code{app$min} and \code{app$max} inclusive).
  \item Each element of \code{val} must be validate by \code{app$prototype}.
        \code{transform.labeled} is passed on.
}
}
\author{
Brad Friedman
}

