% Generated by roxygen2 (4.0.2): do not edit by hand
\name{validate.bool.param.value}
\alias{validate.bool.param.value}
\title{validate.bool.param.value}
\usage{
validate.bool.param.value(app, val)
}
\arguments{
\item{app}{AnalysisPageParam}

\item{val}{Candidate value}
}
\value{
\code{val} unmodified
}
\description{
Validate a boolean-type AnalysisPageParam value
}
\details{
Validate a boolean-type AnalysisPageParam value:
\enumerate{
  \item \code{val} must be length 1
  \item \code{val} must be a logical
}
}
\author{
Brad Friedman
}

