% Generated by roxygen2 (4.0.2): do not edit by hand
\name{validate.combobox.param.value}
\alias{validate.combobox.param.value}
\title{Validate a combobox-type AnalysisPageParam value}
\usage{
validate.combobox.param.value(app, val, transform.labeled = FALSE)
}
\arguments{
\item{app}{AnalysisPageParam}

\item{val}{Candidate value}

\item{transform.labeled}{Logical. See details.}
}
\value{
Candidate value, possibly transformed into list.
}
\description{
Alias for \code{\link{validate.labeled.param.value}}
}

