% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/validate.param.list.R
\name{validate.compound.param.value}
\alias{validate.compound.param.value}
\title{validate.compound.param.value}
\usage{
validate.compound.param.value(app, val, transform.labeled = FALSE)
}
\arguments{
\item{app}{AnalysisPageParam}

\item{val}{Candidate value}

\item{transform.labeled}{Passed through to \code{\link{validate.param.value}} of children. Default: FALSE.}
}
\value{
value, possibly with labeled children transformed
}
\description{
Validate a compound-type AnalysisPageParam value
}
\details{
Validate a compound-type AnalysisPageParam value:
\enumerate{
  \item \code{val} must be a list. If length 0 then it is valid with no further checks.
  \item \code{names(val)} must not have any duplicates.
  \item \code{names(val)} must be a subset of \code{names(app$children)}.
  \item Each of the values in the list must be validated by the corresponding
        child. \code{transform.labeled} is passed on.
}
}
\author{
Brad Friedman
}

