% Generated by roxygen2 (4.0.2): do not edit by hand
\name{validate.file.param.value}
\alias{validate.file.param.value}
\title{validate.file.param.value}
\usage{
validate.file.param.value(app, val)
}
\arguments{
\item{app}{AnalysisPageParam}

\item{val}{Candidate value}
}
\value{
Never returns
}
\description{
Validate a file-type AnalysisPageParam value
}
\details{
Validate a file-type AnalysisPageParam value:

Current all values are invalid and result in an error being thrown.
The reason for this is that the use case I have in mind is to check
values when constructing a URL, and I don't think file-uploads will
be allowed to be URL-encoded. So I can't think what values will be
valid. Once I have another use case where they ought be valid then
I will know what form they should take.
}
\author{
Brad Friedman
}

