% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.param.list.R
\name{validate.labeled.param.value}
\alias{validate.labeled.param.value}
\title{validate.labeled.param.value}
\usage{
validate.labeled.param.value(app, val, transform.labeled = FALSE)
}
\arguments{
\item{app}{AnalysisPageParam}

\item{val}{Candidate value}

\item{transform.labeled}{Logical. See details.}
}
\value{
Candidate value, possibly transformed into list.
}
\description{
Validate a labeled AnalysisPageParam value
}
\details{
Validate a labeled AnalysisPageParam value. Currently the labeled
param types are "combobox" and "select".

Unnamed scalars are OK and named scalars are also OK.

If the scalar is unnamed then its own name will be applied: \code{names(val) <- val}.

If \code{transform.labeled} is set then instead of returning
the candidate value as-is, it is transformed into
\code{list(v=real.value, r=readable.value)}.

If \code{$allow.multiple == TRUE} then \code{length(val) > 1} is OK.
The encoding is simply \code{list(v=real.values, r=readable.values)},
where \code{real.values} and \code{readable.values} are equal-length vectors.
}
\author{
Brad Friedman
}

