% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/validate.param.list.R
\name{validate.param.list}
\alias{validate.param.list}
\title{validate.param.list}
\usage{
validate.param.list(pset, plist, transform.labeled = FALSE)
}
\arguments{
\item{pset}{AnalysisPageParamSet}

\item{plist}{List of parameter values for a subset of the parameters.}

\item{transform.labeled}{If TRUE, then labeled parameters (combobox
and select) will be transformed as necessary so that they have the
\code{list(v=value, r=readable.value)} format.>
If FALSE (default), then they will be left as-is.}
}
\value{
Copy of \code{plist}, possibly transformed.
}
\description{
Validate a list of parameter values for an AnalysisPageParamSet
}
\details{
Validate a list of parameter values for an AnalysisPageParamSet.
}
\author{
Brad Friedman
}

