% Generated by roxygen2 (4.0.2): do not edit by hand
\name{validate.param.value}
\alias{validate.param.value}
\title{validate.param.value}
\usage{
validate.param.value(app, val, transform.labeled = FALSE)
}
\arguments{
\item{app}{AnalysisPageParam}

\item{val}{Candidate value}

\item{transform.labeled}{Logical. Should the parameter value be transformed
if necessary to have the \code{list(v=value, r=readable.value)}
format?}
}
\value{
The candidate value, possibly transformed. Throws error if the value is invalid.
}
\description{
Validate a parameter value for an AnalysisPageParam
}
\details{
This function dispatches to the type-specific validator.

\code{transform.labeled} is passed on only if that validator accepts
such an argument. This should be just combobox and select types,
and indicates that the parameter value should be transformed to have the
\code{list(v=value, r=readable.value)} format.
}
\examples{
sp <- simple.param("foo")
validate.param.value(sp, 3)
}
\author{
Brad Friedman
}

