% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.param.list.R
\name{validate.select.param.value}
\alias{validate.select.param.value}
\title{validate.select.param.value}
\usage{
validate.select.param.value(app, val, transform.labeled = FALSE)
}
\arguments{
\item{app}{AnalysisPageParam}

\item{val}{Candidate value}

\item{transform.labeled}{Logical. See details.}
}
\value{
Candidate value, possibly transformed into list.
}
\description{
Validate a select-type AnalysisPageParam value
}
\details{
Validate a select-type AnalysisPageParam value.

\enumerate{
  \item The value must be a real value among the choices.
  \item The value must validate by \code{validate.labeled.param.value}
}

If \code{transform.labeled} is set then instead of returning
the candidate value as-is, it is transformed into
\code{list(v=real.value, r=readable.value)}.
}
\author{
Brad Friedman
}
