% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/validate.param.list.R
\name{validate.text.param.value}
\alias{validate.text.param.value}
\title{validate.text.param.value}
\usage{
validate.text.param.value(app, val)
}
\arguments{
\item{app}{AnalysisPageParam}

\item{val}{Candidate value}
}
\value{
\code{val} unmodified
}
\description{
Validate a text-type AnalysisPageParam value
}
\details{
Validate a text-type AnalysisPageParam value:
\enumerate{
  \item \code{val} must be a scalar (length-1 atomic)
  \item \code{val} must be unnamed
}
}
\author{
Brad Friedman
}

