% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/error.R
\name{vwc.conditions}
\alias{vwc.conditions}
\title{vwc.conditions}
\usage{
vwc.conditions(vwc, type = "messages")
}
\arguments{
\item{vwc}{AnalysisPageValueWithConditions}

\item{type}{"messages" "warnings" or "error"}
}
\value{
List of condition objects for "messages" or "warnings"
or a single condition object or NULL for "error".
}
\description{
Return condition object(s) for an AnalysisPageValueWithConditions
}
\details{
Return condition object(s) for an AnalysisPageValueWithConditions
}
\examples{
vwc <- tryKeepConditions({message("whatever"); warning("warning message"); 3})
vwc.conditions(vwc, "messages")
vwc.conditions(vwc, "warnings")
vwc.conditions(vwc, "error")
}
\author{
Brad Friedman
}
\seealso{
\code{\link{vwc.error.condition}},
\code{\link{vwc.messages.conditions}},
\code{\link{vwc.warnings.conditions}}
}

