% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{vwc.error.condition}
\alias{vwc.error.condition}
\title{vwc.error.condition}
\usage{
vwc.error.condition(vwc)
}
\arguments{
\item{vwc}{AnalysisPageValueWithCondition}
}
\value{
condition object for error, or NULL
}
\description{
Get condition object for error
}
\details{
Get condition object for error
}
\examples{
vwc <- tryKeepConditions({stop("error!")})
vwc.error.condition(vwc)
}
\author{
Brad Friedman
}
