% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{vwc.error.traceback}
\alias{vwc.error.traceback}
\title{vwc.error.traceback}
\usage{
vwc.error.traceback(vwc)
}
\arguments{
\item{vwc}{AnalysisPageValueWithConditions}
}
\value{
Charvecs, or NULL if there was no error.
(In this it differs from \code{\link{vwc.tracebacks}}, which
 throws an error).
}
\description{
Get traceback for error
}
\details{
Get traceback for error
}
\examples{
f <- function(msg)  {
  stop(msg)
}
vwc <- tryKeepConditions({
  f("foo")
})
vwc.error.traceback(vwc)
}
\author{
Brad Friedman
}
\seealso{
\code{\link{vwc.tracebacks}}
}

