% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{vwc.messages}
\alias{vwc.messages}
\title{vwc.messages}
\usage{
vwc.messages(vwc, type = "messages")
}
\arguments{
\item{vwc}{AnalysisPageValueWithConditions}

\item{type}{Type of conditions. Must be "messages" or "warnings".
Default: messages.}
}
\value{
Charvec of message strings
}
\description{
Return condition Messages for an AnalysisPageValueWithConditions
}
\details{
Return condition Messages for an AnalysisPageValueWithConditions.
}
\examples{
vwc <- tryKeepConditions({message("I've got something to say.")})
vwc.messages(vwc)
}
\author{
Brad Friedman
}

