% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/error.R
\name{vwc.messages.tracebacks}
\alias{vwc.messages.tracebacks}
\title{vwc.messages.tracebacks}
\usage{
vwc.messages.tracebacks(vwc)
}
\arguments{
\item{vwc}{AnalysisPageValueWithConditions}
}
\value{
List of charvecs.
}
\description{
Get list of messages tracebacks
}
\details{
Get list of tracebacks for messages.
}
\examples{
f <- function(msg)  {
  message(msg)
}
vwc <- tryKeepConditions({
  f("foo")
  f("bar")
})
vwc.messages.tracebacks(vwc)
}
\author{
Brad Friedman
}
\seealso{
\code{\link{vwc.tracebacks}}
}

