% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{vwc.n}
\alias{vwc.n}
\title{vwc.n}
\usage{
vwc.n(vwc, type)
}
\arguments{
\item{vwc}{AnalysisPageValueWithConditions}

\item{type}{"messages", "warnings" or "error"}
}
\value{
Number of conditions. (For "error" it can only be 0 or 1, and is equivalent
to calling \code{as.integer(\link{vwc.is.error}())}).
}
\description{
Get number of conditions of a given type for an AnalysisPageValueWithConditions
}
\details{
Get number of conditions of a given type for an AnalysisPageValueWithConditions.
}
\examples{
vwc <- tryKeepConditions({stop("This isn't going to work")})
vwc.n(vwc, "error")
}
\author{
Brad Friedman
}
