% Generated by roxygen2 (4.0.2): do not edit by hand
\name{vwc.n.warnings}
\alias{vwc.n.warnings}
\title{vwc.n.warnings}
\usage{
vwc.n.warnings(vwc)
}
\arguments{
\item{vwc}{AnalysisPageValueWithConditions}
}
\value{
Non-negative Integer
}
\description{
Get number of warnings for an AnalysisPageValueWithConditions
}
\details{
Get number of warnings for an AnalysisPageValueWithConditions
}
\examples{
vwc <- tryKeepConditions({warning("I'm warning you.")})
vwc.n.warnings(vwc)
}
\author{
Brad Friedman
}

