% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/error.R
\name{vwc.tracebacks}
\alias{vwc.tracebacks}
\title{vwc.tracebacks}
\usage{
vwc.tracebacks(vwc, type = "messages")
}
\arguments{
\item{vwc}{AnalysisPageValueWithConditions}

\item{type}{Type of conditions. Must be "messages", "warnings" or "error".
Default: "messages". A (new) error is thrown if type is "error" but
the \code{vwc} is not an error (that is, does not have an error, or
more specifically, \code{!vwc.is.error(vwc)}).}
}
\value{
For "messages" or "warnings" it gives
a list of Charvecs of tracebacks, as built by \code{\link{getTraceback}}.
For "error" it only gives a single charvec, since there is only one error.
}
\description{
Return traceback(s) for an AnalysisPageValueWithConditions
}
\details{
Return traceback(s) for an AnalysisPageValueWithConditions
}
\examples{
f <- function(msg)  {
  warning(msg)
}
vwc <- tryKeepConditions({
  f("foo")
  f("bar")
})
vwc.tracebacks(vwc, "warnings")
}
\author{
Brad Friedman
}

