% Generated by roxygen2 (4.0.2): do not edit by hand
\name{vwc.value}
\alias{vwc.value}
\title{vwc.value}
\usage{
vwc.value(vwc)
}
\arguments{
\item{vwc}{AnalysisPageValueWithConditions}
}
\value{
Value of original evaluated expression, or NULL if an error was thrown.
}
\description{
Get value of any AnalysisPageValueWithConditions
}
\details{
Get value of any AnalysisPageValueWithConditions. If an error was thrown
then the value will be NULL.
}
\examples{
vwc <- tryKeepConditions({3+5})
vwc.value(vwc)
}
\author{
Brad Friedman
}

