% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/error.R
\name{vwc.warnings}
\alias{vwc.warnings}
\title{vwc.warnings}
\usage{
vwc.warnings(vwc)
}
\arguments{
\item{vwc}{AnalysisPageValueWithConditions}
}
\value{
Charvec of warning messages
}
\description{
Access warning messages from AnalysisPageValueWithConditions
}
\details{
Access warning messages from AnalysisPageValueWithConditions
}
\examples{
vwc <- tryKeepConditions({warning("I'm warning you!"); warning("Again")})
vwc.warnings(vwc)
}
\author{
Brad Friedman
}

