% Generated by roxygen2 (4.0.2): do not edit by hand
\name{vwc.warnings.tracebacks}
\alias{vwc.warnings.tracebacks}
\title{vwc.warnings.tracebacks}
\usage{
vwc.warnings.tracebacks(vwc)
}
\arguments{
\item{vwc}{AnalysisPageValueWithConditions}
}
\value{
List of charvecs.
}
\description{
Get list of warnings tracebacks
}
\details{
Get list of tracebacks for warnings.
}
\examples{
f <- function(msg)  {
  warning(msg)
}
vwc <- tryKeepConditions({
  f("foo")
  f("bar")
})
vwc.warnings.tracebacks(vwc)
}
\author{
Brad Friedman
}
\seealso{
\code{\link{vwc.tracebacks}}
}

