% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/client.ip.R
\name{client.ip}
\alias{client.ip}
\title{client.ip}
\usage{
client.ip()
}
\value{
\code{SERVER$remote_ip}
}
\description{
Return client IP address
}
\details{
This is the "X-Forwarded-For" header, if available, and otherwise the "remote_ip" component of the global SERVER variable
}
\examples{
SERVER <<- list(remote_ip = "127.0.0.1")
client.ip()
}
\author{
Brad Friedman
}

