% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom.html.R
\name{default.stylesheets}
\alias{default.stylesheets}
\title{default.stylesheets}
\usage{
default.stylesheets()
}
\value{
character vector
}
\description{
Default stylesheets for HTML headers
}
\details{
Default stylesheets for HTML headers
}
\author{
Brad Friedman
}
