% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static.analysis.page.R
\name{get.APS.outdir}
\alias{get.APS.outdir}
\title{Get current AnalysisPageServer output directory}
\usage{
get.APS.outdir()
}
\value{
Path
}
\description{
Get current AnalysisPageServer output directory
}
\examples{
set.APS.outdir("/some/path")
get.APS.outdir()
reset.APS.outdir()
}
\author{
Brad Friedman
}
\seealso{
\code{\link{set.APS.outdir}}, \code{\link{reset.APS.outdir}}
}

