% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/validate.params.R
\name{known.param.sizes}
\alias{known.param.sizes}
\title{known.param.sizes}
\usage{
known.param.sizes()
}
\value{
Character vector
}
\description{
Get the controlled vocabulary of parameter size words
}
\details{
Returns the controlled vocabulary of parameter size words
}
\author{
Brad Friedman
}

