% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.page.R
\name{make.standard.ids}
\alias{make.standard.ids}
\title{make.standard.ids}
\usage{
make.standard.ids(n, prefix = "Reg")
}
\arguments{
\item{n}{Desired length of vector}

\item{prefix}{String, default "Reg".}
}
\value{
Character vector. Currently just "Reg1", "Reg2", ..., "Regn" (or starting with whatever prefix is).
}
\description{
Make a vector of standardized IDs
}
\details{
Make a vector of standardized IDs.
}
\author{
Brad Friedman - Regular
}
