% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rook.analysis.page.server.R
\name{rook.analysis.page.server.landing.page}
\alias{rook.analysis.page.server.landing.page}
\title{rook.analysis.page.server.landing.page}
\usage{
rook.analysis.page.server.landing.page(baseurl)
}
\arguments{
\item{baseurl}{String. Base URL (typically ending in "/custom/RAPS"), or a list with \code{$url} element.}
}
\value{
String, full URL to landing page.
}
\description{
Return URL for landing page of a Rook AnalysisPageServer
}
\details{
Return URL for landing page of a Rook AnalysisPageServer
}
\examples{
message("See vignette ExamplesServers.html")
}
\author{
Brad Friedman
}

