% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{trig.registry}
\alias{trig.registry}
\title{trig.registry}
\usage{
trig.registry()
}
\value{
AnalysisPageRegistry
}
\description{
Build a toy registry for examples and testing
}
\details{
The toy registry has a sine and a cosine page and the scattergram page.
}
\examples{
tr <- trig.registry()
pages(tr)
}
\author{
Brad Friedman
}

