% Generated by roxygen2 (4.0.2): do not edit by hand
\name{urlDecode}
\alias{urlDecode}
\title{urlDecode a string}
\arguments{
\item{x}{Character vector of strings to urlDecode}
}
\value{
Character vector of same length as \code{x} containing
decoded strings
}
\description{
When a function of this name is available from the global environment
(such as when running under RApache) then that function is used.
Otherwise a pure R implementation is provided.
}
\author{
Brad Friedman
}

