% Generated by roxygen2 (4.0.2): do not edit by hand
\name{vwc.error.condition}
\alias{vwc.error.condition}
\title{vwc.error.condition}
\usage{
vwc.error.condition(vwc)
}
\arguments{
\item{vwc}{AnalysisPageValueWithCondition}
}
\value{
condition object for error, or NULL
}
\description{
Get condition object for error
}
\details{
Get condition object for error
}
\examples{
vwc <- tryKeepConditions({stop("error!")})
vwc.error.condition(vwc)
}
\author{
Brad Friedman
}

