% Generated by roxygen2 (4.0.2): do not edit by hand
\name{vwc.error}
\alias{vwc.error}
\title{vwc.error}
\usage{
vwc.error(vwc)
}
\arguments{
\item{vwc}{AnalysisPageValueWithConditions}
}
\value{
Charvec of warning messages
}
\description{
Access error message from AnalysisPageValueWithConditions
}
\details{
Access error message from AnalysisPageValueWithConditions
}
\examples{
vwc <- tryKeepConditions({stop("This isn't going to work")})
vwc.error(vwc)
}
\author{
Brad Friedman
}

