% Generated by roxygen2 (4.0.2): do not edit by hand
\name{vwc.is.error}
\alias{vwc.is.error}
\title{vwc.is.error}
\usage{
vwc.is.error(vwc)
}
\arguments{
\item{vwc}{AnalysisPageValueWithConditions, as returned by \code{\link{tryKeepConditions}}}
}
\value{
Logical
}
\description{
Predicate to test if an AnalysisPageValueWithConditions had an error
}
\details{
Predicate to test if an AnalysisPageValueWithConditions had an error
}
\examples{
vwc <- tryKeepConditions({3+5})
vwc.is.error(vwc)

vwc <- tryKeepConditions({stop("error!")})
vwc.is.error(vwc)
}
\author{
Brad Friedman
}

