% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{vwc.n.messages}
\alias{vwc.n.messages}
\title{vwc.n.messages}
\usage{
vwc.n.messages(vwc)
}
\arguments{
\item{vwc}{AnalysisPageValueWithConditions}
}
\value{
Non-negative Integer
}
\description{
Get number of messages for an AnalysisPageValueWithConditions
}
\details{
Get number of messages for an AnalysisPageValueWithConditions
}
\examples{
vwc <- tryKeepConditions({message("Message")})
vwc.n.messages(vwc)
}
\author{
Brad Friedman
}
