% Generated by roxygen2 (4.0.2): do not edit by hand
\name{vwc.warnings.conditions}
\alias{vwc.warnings.conditions}
\title{vwc.warnings.conditions}
\usage{
vwc.warnings.conditions(vwc)
}
\arguments{
\item{vwc}{AnalysisPageValueWithCondition}
}
\value{
List of condition object for warnings (might be of length 0 if
no warnings were thrown).
}
\description{
Get condition object for warnings
}
\details{
Get condition object for warnings
}
\examples{
vwc <- tryKeepConditions({warning("I'm warning you!")})
vwc.warnings.conditions(vwc)
}
\author{
Brad Friedman
}

